Add function to wait for the OCP to be ready
When "lab" script is run for the first time in the fresh lab environment
it may happen that SSL certificates in the OCP cluster will not be
refreshed yet thus preparation of the bfx scenario may fail.
To avoid that, this patch adds small function to wait for the OCP
cluster and API to be ready and working properly before moving on with
other configuration for the bfx scenario.

Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>