Add info that script is waiting for the CR to be ready.
One of the things which lab script is doing is to make sure that
OpenStackContolPlane CR has properly configured nicMappings. This
requires some time to be finished and script is simply waiting for the
CR to be ready. Now there is small message added around that waiting
time so that user will know what happens there instead of having
impression that script is stucked somehow.

Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>